+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - 🟥 builder fns now require `size` as first param
   - 🟥 builder `.tabs` renamed to `.dock` for more consistent terminology
   - 🟥 builder `.cell` renamed to `.horizontal` for more consistent terminology
+- 🔶 `Layout.version` schema bump (will reset to default)
 - 🍏 added react compat via `reactIntegration`, see readme
 - 🍏 improved behaviors and bug fixes
 - 🍏 added `--gutter-size: 0.7em;` css variable